+++ packages/example/src/components/common/sdk-forms/price-form.tsx
@@ -1,7 +1,5 @@
 import React, { useState } from "react"
-import {
-	Blockchain,
-} from "@rarible/api-client"
+import { Blockchain } from "@rarible/api-client"
 import { Grid, MenuItem, Stack } from "@mui/material"
 import { FormTextInput } from "../form/form-text-input"
 import { UseFormReturn } from "react-hook-form"
@@ -11,7 +9,8 @@ import { ContractAddress, toContractAddress } from "@rarible/types"
 
 interface IPriceFormProps {
 	form: UseFormReturn
-	currencyOptions: CurrencyOption[]
+	currencyOptions?: CurrencyOption[],
+	max?: string | number
 }
 
 function getCurrencyOptionValue(option: CurrencyOption): string {
@@ -23,11 +22,11 @@ export function parseCurrencyType(value: string): {
 	type: CurrencyOption["type"],
 	contract: ContractAddress | undefined,
 } {
-	const [blockchain,type, contract] = value.split("::")
+	const [blockchain, type, contract] = value.split("::")
 	return {
 		blockchain: blockchain as Blockchain,
 		type: type as CurrencyOption["type"],
-		contract: contract ? toContractAddress(contract) : undefined
+		contract: contract ? toContractAddress(contract) : undefined,
 	}
 }
 
@@ -36,8 +35,14 @@ function getCurrencyOptionByValue(value: string, currencyOptions: CurrencyOption
 }
 
 export function PriceForm(props: IPriceFormProps) {
-	const { form, currencyOptions } = props
-	const [currencyType, setCurrencyType] = useState(getCurrencyOptionValue(currencyOptions[0]))
+	const {
+		form,
+		currencyOptions: currencyOptionsOriginal,
+		max,
+	} = props
+	const currencyOptions = currencyOptionsOriginal || []
+	const isEmptyCurrency = currencyOptions !== undefined
+	const [currencyType, setCurrencyType] = useState(currencyOptions[0] && getCurrencyOptionValue(currencyOptions[0]))
 	const selectedOption = getCurrencyOptionByValue(currencyType, currencyOptions)
 
 	return <Stack spacing={2}>
@@ -45,40 +50,49 @@ export function PriceForm(props: IPriceFormProps) {
 			<Grid item xs={8}>
 				<FormTextInput
 					type="number"
-					inputProps={{ min: 0, step: "any" }}
+					inputProps={{
+						min: 0,
+						step: "any",
+						max: max || undefined,
+					}}
 					form={form}
 					options={{
 						min: 0,
+						max: max || undefined,
 					}}
 					name="price"
 					label="Price"
 				/>
 			</Grid>
-			<Grid item xs={4}>
-				<FormSelect
-					form={form}
-					value={currencyType}
-					onChange={(e) => {
-						setCurrencyType(e.target.value)
-						const selectedOption = getCurrencyOptionByValue(e.target.value, currencyOptions)
-						form.setValue("contract", selectedOption?.type === "TOKEN" ? selectedOption.contract ?? "" : "")
-					}}
-					name="currencyType"
-					label="Currency"
-				>
-					{
-						currencyOptions.map((option, index) => {
-							const value = getCurrencyOptionValue(option)
-							return <MenuItem key={value} value={value}>
-								{option.label}
-							</MenuItem>
-						})
-					}
-				</FormSelect>
-			</Grid>
+			{
+				isEmptyCurrency && currencyOptions.length ?
+					<Grid item xs={4}>
+						<FormSelect
+							form={form}
+							value={currencyType}
+							onChange={(e) => {
+								setCurrencyType(e.target.value)
+								const selectedOption = getCurrencyOptionByValue(e.target.value, currencyOptions)
+								form.setValue("contract", selectedOption?.type === "TOKEN" ? selectedOption.contract ?? "" : "")
+							}}
+							name="currencyType"
+							label="Currency"
+						>
+							{
+								currencyOptions.map((option, index) => {
+									const value = getCurrencyOptionValue(option)
+									return <MenuItem key={value} value={value}>
+										{option.label}
+									</MenuItem>
+								})
+							}
+						</FormSelect>
+					</Grid>
+					: null
+			}
 		</Grid>
 		{
-			selectedOption?.type === "TOKEN" ?
+			selectedOption?.type === "TOKEN" && isEmptyCurrency ?
 				<FormTextInput
 					type="text"
 					form={form}
@@ -87,7 +101,7 @@ export function PriceForm(props: IPriceFormProps) {
 					name="contract"
 					label="Contract"
 				/>
-			: null
+				: null
 		}
 	</Stack>
 }