+++ packages/sdk/src/sdk-blockchains/ethereum/bid.ts
@@ -239,7 +239,7 @@ export class EthereumBid {
 						originFeesSum
 					)
 					await this.convertCurrency(value)
-					console.log("after convert")
+					// console.log("after convert")
 				}
 				return request
 			},